sync clients have their own place
