+++ settings/templates/settings/personal/sync-clients.php
@@ -1,0 +1,59 @@
+<?php
+/**
+ * @copyright Copyright (c) 2017 Arthur Schiwon <blizzz@arthur-schiwon.de>
+ *
+ * @author Arthur Schiwon <blizzz@arthur-schiwon.de>
+ *
+ * @license GNU AGPL version 3 or any later version
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
+/** @var array $_ */
+
+?>
+
+<div id="clientsbox" class="section clientsbox">
+	<h2><?php p($l->t('Get the apps to sync your files'));?></h2>
+	<a href="<?php p($_['clients']['desktop']); ?>" rel="noreferrer" target="_blank">
+		<img src="<?php print_unescaped(image_path('core', 'desktopapp.svg')); ?>"
+			 alt="<?php p($l->t('Desktop client'));?>" />
+	</a>
+	<a href="<?php p($_['clients']['android']); ?>" rel="noreferrer" target="_blank">
+		<img src="<?php print_unescaped(image_path('core', 'googleplay.png')); ?>"
+			 alt="<?php p($l->t('Android app'));?>" />
+	</a>
+	<a href="<?php p($_['clients']['ios']); ?>" rel="noreferrer" target="_blank">
+		<img src="<?php print_unescaped(image_path('core', 'appstore.svg')); ?>"
+			 alt="<?php p($l->t('iOS app'));?>" />
+	</a>
+
+	<p>
+		<?php print_unescaped(str_replace(
+			[
+				'{contributeopen}',
+				'{linkclose}',
+			],
+			[
+				'<a href="https://nextcloud.com/contribute" target="_blank" rel="noreferrer">',
+				'</a>',
+			],
+			$l->t('If you want to support the project {contributeopen}join development{linkclose} or {contributeopen}spread the word{linkclose}!'))); ?>
+	</p>
+
+	<?php if(OC_APP::isEnabled('firstrunwizard')) {?>
+		<p><a class="button" href="#" id="showWizard"><?php p($l->t('Show First Run Wizard again'));?></a></p>
+	<?php }?>
+</div>
